test: add class BookingServicesTest
